requirements.txt
pandas~=2.2.1
scikit-learn~=1.4.1.post1
tensorflow~=2.15.0
keras~=2.15.0
pillow
boto3